Removed all schemas of the theme
