+++ src/components/Card/index.js
@@ -3,7 +3,7 @@
 import React, { type Node } from 'react';
 
 import createSheet from '../../styles/create-sheet';
-import { type Theme } from '../../theme/schema';
+import { type Theme } from '../../theme/types';
 
 import Actions from './Actions';
 import Content from './Content';
