+++ src/components/Divider/index.js
@@ -3,7 +3,7 @@
 import React from 'react';
 
 import createSheet from '../../styles/create-sheet';
-import { type Theme } from '../../theme/schema';
+import { type Theme } from '../../theme/types';
 
 type Props = {
   vertical: boolean,
