+++ src/components/ExpansionPanel/Actions.jsx
@@ -3,7 +3,7 @@
 import React, { type Node } from 'react';
 
 import createSheet from '../../styles/create-sheet';
-import { type Theme } from '../../theme/schema';
+import { type Theme } from '../../theme/types';
 
 type Props = {
   children: Node,
