+++ src/components/Tabs/Sheet.js
@@ -2,7 +2,7 @@
 
 import createSheet from '../../styles/create-sheet';
 import { getActiveColor } from '../../theme/utils';
-import { type Theme } from '../../theme/schema';
+import { type Theme } from '../../theme/types';
 
 export type Data = {
   transform: string | null,
