+++ src/components/TextField/Input.jsx
@@ -4,7 +4,7 @@ import React, { type Node } from 'react';
 
 import createSheet from '../../styles/create-sheet';
 import Typography from '../Typography';
-import { type Theme } from '../../theme/schema';
+import { type Theme } from '../../theme/types';
 import { getActiveColor } from '../../theme/utils';
 
 type Props = {
