+++ src/components/TextField/Underline.jsx
@@ -3,7 +3,7 @@
 import React from 'react';
 
 import createSheet from '../../styles/create-sheet';
-import { type Theme } from '../../theme/schema';
+import { type Theme } from '../../theme/types';
 import { getActiveColor } from '../../theme/utils';
 
 type Props = {
