+++ src/theme/z-indexes.js
@@ -1,10 +1,5 @@
 // @flow strict
 
-import {
-  number,
-  shape,
-} from 'prop-types';
-
 type ZIndexes = {
   appBar: number,
   fab: number,
@@ -13,15 +8,6 @@ type ZIndexes = {
   dialog: number,
   drawer: number,
 };
-
-const schema = shape({
-  appBar: number.isRequired,
-  fab: number.isRequired,
-  snackbar: number.isRequired,
-  backdrop: number.isRequired,
-  dialog: number.isRequired,
-  drawer: number.isRequired,
-}).isRequired;
 
 const defaultZIndexes: ZIndexes = {
   appBar: 50,
@@ -34,7 +20,4 @@ const defaultZIndexes: ZIndexes = {
 
 export type { ZIndexes };
 
-export {
-  schema,
-  defaultZIndexes,
-};
+export { defaultZIndexes };
